# Refactored Position of Regexp