I moved regexps as package valuables for decreasing memory allocations.

benchmark code (`logmon_test.go`, same folder w/ `logmon.go`)
```go
package main

import (
	"testing"
)

func Benchmark_parseConf(b *testing.B) {
	for i := 0; i < b.N; i++ {
			parseConf(`:/var/log/system.log\n(ERROR)\necho -e 'ERROR\n<%%%%>\n'`)
	}
}
```

before
```bash
$ go test -bench . -benchmem
goos: darwin
goarch: amd64
pkg: github.com/MakoTano/logmon-go
Benchmark_parseConf-4   	   50000	     27605 ns/op	   55184 B/op	     198 allocs/op
PASS
ok  	github.com/MakoTano/logmon-go	1.683s
```

this PR
```bash
$ go test -bench . -benchmem
goos: darwin
goarch: amd64
pkg: github.com/MakoTano/logmon-go
Benchmark_parseConf-4   	  500000	      2888 ns/op	     192 B/op	       5 allocs/op
PASS
ok  	github.com/MakoTano/logmon-go	1.482s
```

thank you for nice OSS, I've using this in our production. I love it.